feat: restructure src
