--- src/Assets/Styling/index.scss
@@ -1,47 +1,0 @@
-// Imports
-@import 'colors'; // og colors
-@import 'reset'; // minimal reset ruleset
-@import 'mixins';
-
-//BREAK-POINTS
-$screen-xs: 23.125rem; // XSmall. Window width >=370px
-$screen-sm: 36rem; // Small. Window width >=576px
-$screen-md: 53.25rem; // Medium. Window width >=852px
-$screen-lg: 60.5rem; // Large. Window width >=968px
-$screen-xl: 67.5rem; // XLarge. Window width >=1080px
-
-// GLOBAL
-html {
-  scroll-behavior: smooth;
-}
-
-body {
-  background-color: $colors-dark-background;
-}
-
-::-webkit-scrollbar {
-  width: 8px;
-  &-thumb {
-    background: $colors-dark-textSubtle;
-    border-radius: 0.5rem;
-  }
-  &-track {
-    box-shadow: inset 0 0 0.313rem $colors-dark-input;
-    border-radius: 0.625rem;
-  }
-}
-
-// Component's sccs file import
-@import './../Components/PhishingBanner/PhishingBanner';
-@import './../Components/Header/Header';
-@import './../Components/Header/NavBar/NavBar';
-@import './../Components/BottomNav/BottomNav';
-
-// App component wrapper
-.content-wrapper {
-  position: relative;
-  width: 100%;
-  display: grid;
-  grid-template-rows: auto 1fr;
-  min-height: auto;
-}
