+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as Header } from './header/Header';
+export { default as BottomNav } from './bottomNav/BottomNav';
