+++ src/components/index.scss
@@ -1,0 +1,4 @@
+@import './phishingBanner/PhishingBanner';
+@import './header/Header';
+@import './header/NavBar';
+@import './bottomNav/BottomNav';
