+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './Assets/Styling/index.scss';
+import './index.scss';
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('App'));
