+++ src/index.scss
@@ -1,0 +1,33 @@
+// Imports
+@import './mixins/index.scss';
+@import './components/index.scss';
+
+// GLOBAL
+html {
+  scroll-behavior: smooth;
+}
+
+body {
+  background-color: $colors-dark-background;
+}
+
+::-webkit-scrollbar {
+  width: 8px;
+  &-thumb {
+    background: $colors-dark-textSubtle;
+    border-radius: 0.5rem;
+  }
+  &-track {
+    box-shadow: inset 0 0 0.313rem $colors-dark-input;
+    border-radius: 0.625rem;
+  }
+}
+
+// App component wrapper
+.content-wrapper {
+  position: relative;
+  width: 100%;
+  display: grid;
+  grid-template-rows: auto 1fr;
+  min-height: auto;
+}
