+++ src/mixins/_mixins.scss
@@ -1,5 +1,12 @@
 // MIXINS
 
+//BREAK-POINTS
+$screen-xs: 23.125rem; // XSmall. Window width >=370px
+$screen-sm: 36rem; // Small. Window width >=576px
+$screen-md: 53.25rem; // Medium. Window width >=852px
+$screen-lg: 60.5rem; // Large. Window width >=968px
+$screen-xl: 67.5rem; // XLarge. Window width >=1080px
+
 // for Break-Points
 // XSmall devices
 @mixin xs {
