+++ src/mixins/index.scss
@@ -1,0 +1,3 @@
+@import 'reset';
+@import 'mixins';
+@import 'colors';
